refactor: split progress logging from context
The progress logging does not have much to do with the rest of the
Context class. Move it elsewhere so that each class is more focused on a
single responsibility.
